+++ app/helpers/application_helper.rb
@@ -1,2 +1,19 @@
 module ApplicationHelper
+
+    def mac?
+      !!(ua =~ /Mac OS X/)
+    end
+
+    def windows?
+      !!(ua =~ /Windows/)
+    end
+
+    def linux?
+      !!(ua =~ /Linux/)
+    end
+
+    def ua
+      request.env['HTTP_USER_AGENT']
+    end
+
 end