+++ config/routes.rb
@@ -5,9 +5,14 @@ WwwSleepdataOrg::Application.routes.draw do
   resources :users
 
   get 'welcome/index'
-  get '/about' => "welcome#about", as: :about
-  get '/downloads' => "welcome#downloads", as: :downloads
-  get '/downloads/manifest' => "welcome#manifest", as: :manifest
+  get '/about' => 'welcome#about', as: :about
+  get '/downloads' => 'welcome#downloads', as: :downloads
+  get '/downloads/manifest' => 'welcome#manifest', as: :manifest
+
+  get '/tools' => 'welcome#wget', as: :tools
+  get '/tools/wget' => 'welcome#wget', as: :wget
+  get '/tools/wget/windows' => 'welcome#wget_windows', as: :wget_windows
+  get '/tools/wget/src' => 'welcome#wget_src', as: :wget_src
 
   root to: 'welcome#index'