+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Initial prototype of the National Sleep Research Resource splash page
 - Skeleton files to initialize Rails application with testing framework and continuous integration
 - Added the We Care Dataset to prototype bulk file downloads
+- Added Windows and Mac OS X/Linux instructions for installing GNU Wget